+++ src/Main.hs
@@ -1,6 +1,6 @@
 module Main where
 
-import Control.Lens (view, (^.), to)
+import Control.Lens ((^.), to, (.~), (&))
 import Control.Monad (unless)
 import Control.Monad.Reader (runReaderT)
 import Control.Monad.Random (getRandom)
@@ -18,7 +18,8 @@ import EVM.ABI (AbiValue(AbiAddress))
 import Echidna.ABI
 import Echidna.Config
 import Echidna.Solidity
-import Echidna.Campaign
+import Echidna.Types.Campaign
+import Echidna.Campaign (isSuccess)
 import Echidna.UI
 import Echidna.Transaction
 
@@ -29,6 +30,7 @@ data Options = Options
   { filePath         :: NE.NonEmpty FilePath
   , selectedContract :: Maybe String
   , configFilepath   :: Maybe FilePath
+  , outputFormat     :: Maybe OutputFormat
   }
 
 options :: Parser Options
@@ -40,29 +42,41 @@ options = Options <$> (NE.fromList <$> some (argument str (metavar "FILES"
                   <*> optional (option str $ long "config"
                         <> metavar "CONFIG"
                         <> help "Config file")
+                  <*> optional (option auto $ long "format"
+                        <> metavar "FORMAT"
+                        <> help "Output format: json, text, none. Disables interactive UI")
 
 versionOption :: Parser (a -> a)
 versionOption = infoOption
                   ("Echidna " ++ showVersion version)
                   (long "version" <> help "Show version")
 
-opts :: ParserInfo Options
-opts = info (helper <*> versionOption <*> options) $ fullDesc
+optsParser :: ParserInfo Options
+optsParser = info (helper <*> versionOption <*> options) $ fullDesc
   <> progDesc "EVM property-based testing framework"
   <> header "Echidna"
 
 main :: IO ()
-main = do Options f c conf <- execParser opts
-          g   <- getRandom
-          EConfigWithUsage cfg ks _ <- maybe (pure (EConfigWithUsage defaultConfig mempty mempty)) parseConfig conf
-          unless (cfg ^. sConf . quiet) $ mapM_ (hPutStrLn stderr . ("Warning: unused option: " ++) . unpack) ks
-          let cd = corpusDir $ view cConf cfg
-          txs <- loadTxs cd
-          cpg <- flip runReaderT cfg $ do
-            cs       <- Echidna.Solidity.contracts f
-            ads      <- addresses
-            (v,w,ts) <- loadSpecified (pack <$> c) cs >>= prepareForTest
-            let ads' = AbiAddress <$> v ^. env . EVM.contracts . to keys
-            ui v w ts (Just $ mkGenDict (dictFreq $ view cConf cfg) (extractConstants cs ++ NE.toList ads ++ ads') [] g (returnTypes cs)) txs
-          saveTxs cd (map snd $ DS.toList $ view corpus cpg)
-          if not . isSuccess $ cpg then exitWith $ ExitFailure 1 else exitSuccess
+main = do
+  opts@(Options f c conf _) <- execParser optsParser
+  g <- getRandom
+  EConfigWithUsage loadedCfg ks _ <- maybe (pure (EConfigWithUsage defaultConfig mempty mempty)) parseConfig conf
+  let cfg = overrideConfig loadedCfg opts
+  unless (cfg ^. sConf . quiet) $ mapM_ (hPutStrLn stderr . ("Warning: unused option: " ++) . unpack) ks
+  let cd = cfg ^. cConf . corpusDir
+      df = cfg ^. cConf . dictFreq
+  txs <- loadTxs cd
+  cpg <- flip runReaderT cfg $ do
+    cs       <- Echidna.Solidity.contracts f
+    ads      <- addresses
+    (v,w,ts) <- loadSpecified (pack <$> c) cs >>= prepareForTest
+    let ads' = AbiAddress <$> v ^. env . EVM.contracts . to keys
+    ui v w ts (Just $ mkGenDict df (extractConstants cs ++ NE.toList ads ++ ads') [] g (returnTypes cs)) txs
+  saveTxs cd (snd <$> DS.toList (cpg ^. corpus))
+  if not . isSuccess $ cpg then exitWith $ ExitFailure 1 else exitSuccess
+  where overrideConfig cfg (Options _ _ _ fmt) =
+          case maybe (cfg ^. uConf . operationMode) NonInteractive fmt of
+               Interactive -> cfg
+               NonInteractive Text -> cfg & uConf . operationMode .~ NonInteractive Text
+               nonInteractive -> cfg & uConf . operationMode .~ nonInteractive
+                                     & sConf . quiet .~ True